Total Month Attendance and Last Login & Logout and Logout of the app

+++ app/Http/Controllers/UserController.php
@@ -33,14 +33,22 @@ class UserController extends Controller
 
     function deleteUser($id){
         $deleteUser=User::find($id);
-        $deleteUser->delete();
-        return "User deleted";
+        if($deleteUser){
+             $deleteUser->delete();
+             return "User deleted";
+        }
+       return response()->json(['message'=>'User Not Found'],404);
+       
     }
 
-    function updateUser($id){
+    function updateUser(Request $request,$id){
         $user = User::find($id);
-        $user->Age = "20";
+        if(!$user){
+            return response()->json(['message'=>'User Not Found'],404);
+        }
+        $allRequestData = $request->all();
+        $user->update($allRequestData);
         $user->save();
-        return $user;
+        return response()->json(['message'=>'User Updated successfully'],200);
     }
 }
